feat(scrapers): add new "scrapers" module with first scraper for Bank Yahav (#495)
Add a new `scrapers` package intended to house the scrapers that harvest
data
from bank accounts, credit card companies, and other financial
institutions
and push them into the database (via the GraphQL API) for the user.

This change also adds the first scraper: Bank Yahav in Israel.

In addition, the following changes have been made as well:

### ACME Job (sample tenants job):
- Populate sample tenants YAML via Helm values instead of tenant YAML
files hard-coded inside the Helm chart files. This both removes the need
to ship these files in the chart and also adds flexibility for local
sample tenants in an easier-to-manage manner.
- Fixed a bug in the sample tenant population that has no scrapers or no
accounts.
- Add Helm hooks `post-delete` and `post-rollback` to ensure they get
deleted after uninstallation/rollback
- Add `activeDeadlineSeconds` to ensure the jobs never live more than 1
minute

### Dashboard:
- Upgrade the balance chart (`LineChart` to `LineChartPro`.)
- Allow zooming in and out on the accounts balance chart.
- Add optional `endDate` parameter to account balance query.

### Rates Job:
- Restore the `production` target to the Dockerfile (used by the CI
workflow.)
- Add Helm hooks `post-delete` and `post-rollback` to ensure they get
deleted after uninstallation/rollback
- Add `activeDeadlineSeconds` to ensure the jobs never live more than 1
minute

### Scrapers Page (NEW):
- Add a new page listing all scrapers defined in the tenant
- Page supports editing a scraper and creating a new scraper

### Server:
- Refactor the GraphQL schema
- Refactor the Data Layer for better separation of concerns.

### Database:
- Ensure all SQL constraints are named, in preparation of PostgreSQL
errors parsing which can utilize these names to determine the error
message to return to the user.
- Revert usage of composite primary keys back to a single `id` primary
key, coupled with `UNIQUE` & `FOREIGN KEY` constraints.
- Create the `v_accounts` accounts view which simplifies queries in the
server by coding relevant joins in the PostgreSQL view itself.
- Ensure all `display_name` columns have proper constraints verifying
that the value is not empty
- Restrict tenant ID to 10 characters or less, since tenant IDs are also
used as subdomains.
- Add the `sequence` column to the transaction table and associated
models to reflect transaction ordering within the same date. This is
needed for some transaction data sources that do not provide time-of-day
of the transaction (just dates) and need a way to sort the transactions
within a single day.
- Move database migrations from an init container in the server to a
standalone job that runs on Helm package installation and upgrades.

Dependencies:
- Upgrade `@mui` dependencies to version 8.5.2

Fixes #347